APIs/requirements.txt
Flask==3.1.1
requests
flask_mysqldb
transbank-sdk
flask-cors